refactor(web): split EpisodeChat.tsx (#665)
apps/web/src/components/EpisodeChat.tsx was 531 LOC. Extract:

- EpisodeChatExports.ts: stripHtml, downloadFile, ConversationMeta,
  pairMessages, generateConversationMarkdown,
  generateConversationText.
- EpisodeChatMessage.tsx: the Message type, MessageBubble component,
  scrollToTime / handleCitationClick handlers for click-to-seek on
  citations and source links.

The main file (EpisodeChat.tsx, now 366 LOC) keeps the SSE streaming
state machine, model hydration/persistence, and the panel chrome.
Pure code motion, no behavior changes.
