+++ CHANGELOG.md
@@ -21,6 +21,7 @@ fresh empty `Unreleased` is left at the top.
 ## Unreleased
 
 ### Internal
+- `apps/web/src/components/EpisodeChat.tsx` was 531 LOC; the Markdown/plain-text conversation export helpers move into `EpisodeChatExports.ts` and the assistant/user message bubble + citation handling move into `EpisodeChatMessage.tsx`. The main component now owns just the SSE streaming, model selection, and panel chrome. Pure code motion, no behavior changes. ([#665](https://github.com/brlauuu/podlog/issues/665))
 - `apps/web/src/app/feeds/page.tsx` was 519 LOC; the two Add Feed dialog steps (URL + mode form, and the episode-selection list used by both selective-add and the issue-#487 "add more" flow) move into colocated subcomponents under `apps/web/src/app/feeds/_components/`, and the fetcher helpers + types move into `_lib/`. Pure code motion, no behavior change. ([#664](https://github.com/brlauuu/podlog/issues/664))
 - `apps/web/src/components/RemoteInferenceSectionParts.tsx` was 645 LOC; split along its natural seams — pipeline-step descriptors and helpers into `RemoteInferencePipelineSteps.ts`, the step-card UI cluster (cards + estimates explainer + API-key warning dialog) into `RemoteInferencePipelineCards.tsx`, the Fireworks provider intro and API-key entry into `RemoteInferenceFireworksParts.tsx`, and the pyannote-cloud intro / API-key entry into `RemoteInferencePyannoteParts.tsx`. The original module is now a thin re-export barrel so existing imports keep working. ([#663](https://github.com/brlauuu/podlog/issues/663))
 - `apps/pipeline/app/services/meta_analysis.py` was 565 LOC; split out the per-feed / per-episode / per-speaker / timeline / coverage builders into `meta_analysis_aggregations.py` so the orchestrator (stale helpers, `compute_snapshot`, `upsert_snapshot`, `recompute_and_store`) lives on its own. All previously-importable helpers are still importable from `meta_analysis` for back-compat with existing call sites and tests. ([#662](https://github.com/brlauuu/podlog/issues/662))
