+++ apps/web/src/components/EpisodeChat.tsx
@@ -10,23 +10,21 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { useAudioPlayer } from "@/components/AudioPlayerContext";
-import { MarkdownAnswer, type Source } from "@/lib/citations";
-import { formatTimestamp } from "@/lib/timestamp";
-import { formatDateTime } from "@/lib/dateFormat";
 import { sanitizeFilename } from "@/lib/filename";
 import {
   RAG_MODELS,
   DEFAULT_RAG_MODEL,
   formatModelOption,
 } from "@/lib/rag-models";
+import MessageBubble, { type Message } from "./EpisodeChatMessage";
+import {
+  type ConversationMeta,
+  downloadFile,
+  generateConversationMarkdown,
+  generateConversationText,
+} from "./EpisodeChatExports";
 
 type StreamStatus = "idle" | "connecting" | "streaming" | "done" | "error";
-
-interface Message {
-  role: "user" | "assistant";
-  content: string;
-  sources?: Source[];
-}
 
 // Issue #699: cap history sent to the backend (mirrors the pipeline's own
 // MAX_HISTORY_MESSAGES). 8 messages = 4 prior Q&A pairs.
@@ -39,119 +37,12 @@ interface EpisodeChatProps {
   episodeDescription?: string | null;
 }
 
-function stripHtml(html: string): string {
-  return html.replace(/<[^>]*>/g, "").replace(/\s+/g, " ").trim();
-}
-
-function downloadFile(content: string, filename: string, mimeType: string) {
-  const blob = new Blob([content], { type: mimeType });
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement("a");
-  a.href = url;
-  a.download = filename;
-  document.body.appendChild(a);
-  a.click();
-  document.body.removeChild(a);
-  URL.revokeObjectURL(url);
-}
-
-interface ConversationMeta {
-  feedTitle?: string | null;
-  episodeTitle: string;
-  episodeDescription?: string | null;
-}
-
-function pairMessages(messages: Message[]): { question: string; answer: string; sources: Source[] }[] {
-  const pairs: { question: string; answer: string; sources: Source[] }[] = [];
-  for (let i = 0; i < messages.length; i++) {
-    const m = messages[i];
-    if (m.role !== "user") continue;
-    const next = messages[i + 1];
-    if (next && next.role === "assistant" && next.content) {
-      pairs.push({ question: m.content, answer: next.content, sources: next.sources ?? [] });
-    }
-  }
-  return pairs;
-}
-
-function generateConversationMarkdown(meta: ConversationMeta, messages: Message[]): string {
-  const lines: string[] = [];
-  lines.push(`# ${meta.feedTitle ?? "Podcast"}`);
-  lines.push("");
-  lines.push(`## ${meta.episodeTitle}`);
-  lines.push("");
-  if (meta.episodeDescription) {
-    lines.push(stripHtml(meta.episodeDescription));
-    lines.push("");
-  }
-  lines.push(`*Exported: ${formatDateTime(new Date())}*`);
-  lines.push("");
-  lines.push("---");
-  lines.push("");
-
-  const pairs = pairMessages(messages);
-  pairs.forEach((pair, idx) => {
-    lines.push(`### Q${idx + 1}: ${pair.question}`);
-    lines.push("");
-    lines.push(pair.answer);
-    lines.push("");
-    if (pair.sources.length > 0) {
-      lines.push(`**Sources (${pair.sources.length}):**`);
-      lines.push("");
-      for (const s of pair.sources) {
-        const speaker = s.speaker_label ? ` **${s.speaker_label}**` : "";
-        lines.push(`- \`${s.timestamp ?? formatTimestamp(s.start_time)}\`${speaker} — ${s.text}`);
-      }
-      lines.push("");
-    }
-    lines.push("---");
-    lines.push("");
-  });
-
-  lines.push("*Generated by Podlog*");
-  return lines.join("\n");
-}
-
-function generateConversationText(meta: ConversationMeta, messages: Message[]): string {
-  const lines: string[] = [];
-  lines.push(meta.feedTitle ?? "Podcast");
-  lines.push("=".repeat((meta.feedTitle ?? "Podcast").length));
-  lines.push("");
-  lines.push(meta.episodeTitle);
-  lines.push("-".repeat(meta.episodeTitle.length));
-  lines.push("");
-  if (meta.episodeDescription) {
-    lines.push(stripHtml(meta.episodeDescription));
-    lines.push("");
-  }
-  lines.push(`Exported: ${formatDateTime(new Date())}`);
-  lines.push("");
-  lines.push("========================================");
-  lines.push("");
-
-  const pairs = pairMessages(messages);
-  pairs.forEach((pair, idx) => {
-    lines.push(`Q${idx + 1}: ${pair.question}`);
-    lines.push("");
-    lines.push(`A${idx + 1}: ${pair.answer}`);
-    lines.push("");
-    if (pair.sources.length > 0) {
-      lines.push(`Sources (${pair.sources.length}):`);
-      for (const s of pair.sources) {
-        const speaker = s.speaker_label ? ` ${s.speaker_label}:` : "";
-        lines.push(`  [${s.timestamp ?? formatTimestamp(s.start_time)}]${speaker} ${s.text}`);
-      }
-      lines.push("");
-    }
-    lines.push("----------------------------------------");
-    lines.push("");
-  });
-
-  lines.push("Generated by Podlog");
-  return lines.join("\n");
-}
-
-export default function EpisodeChat({ episodeId, episodeTitle, feedTitle, episodeDescription }: EpisodeChatProps) {
+export default function EpisodeChat({
+  episodeId,
+  episodeTitle,
+  feedTitle,
+  episodeDescription,
+}: EpisodeChatProps) {
   const { state: playerState } = useAudioPlayer();
   const playerVisible = !!playerState.src;
   const [isOpen, setIsOpen] = useState(false);
@@ -326,11 +217,13 @@ export default function EpisodeChat({ episodeId, episodeTitle, feedTitle, episod
         setErrorMsg("Connection failed. Is the pipeline running?");
       }
     },
-    [input, episodeId, model]
+    [input, episodeId, model, messages]
   );
 
   const isStreaming = status === "connecting" || status === "streaming";
-  const hasCompletedExchange = messages.some((m) => m.role === "assistant" && m.content.length > 0);
+  const hasCompletedExchange = messages.some(
+    (m) => m.role === "assistant" && m.content.length > 0,
+  );
 
   const handleExport = useCallback(
     (format: "markdown" | "text") => {
@@ -437,7 +330,11 @@ export default function EpisodeChat({ episodeId, episodeTitle, feedTitle, episod
         )}
 
         {messages.map((msg, i) => (
-          <MessageBubble key={i} message={msg} isStreaming={isStreaming && i === messages.length - 1} />
+          <MessageBubble
+            key={i}
+            message={msg}
+            isStreaming={isStreaming && i === messages.length - 1}
+          />
         ))}
 
         {status === "error" && errorMsg && (
@@ -464,68 +361,6 @@ export default function EpisodeChat({ episodeId, episodeTitle, feedTitle, episod
           {isStreaming ? <Loader2 size={14} className="animate-spin" /> : <Send size={14} />}
         </Button>
       </form>
-    </div>
-  );
-}
-
-function scrollToTime(secs: number) {
-  window.dispatchEvent(
-    new CustomEvent("podlog:scroll-to-time", { detail: { secs } }),
-  );
-}
-
-function handleCitationClick(_episodeId: string, seconds: number) {
-  scrollToTime(seconds);
-}
-
-function MessageBubble({ message, isStreaming }: { message: Message; isStreaming: boolean }) {
-  if (message.role === "user") {
-    return (
-      <div className="flex justify-end">
-        <div className="bg-primary text-primary-foreground rounded-lg px-3 py-2 text-sm max-w-[85%]">
-          {message.content}
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex justify-start">
-      <div className="bg-muted rounded-lg px-3 py-2 text-sm max-w-[85%] space-y-2">
-        {message.content ? (
-          <MarkdownAnswer
-            text={message.content}
-            sources={message.sources ?? []}
-            onCitationClick={handleCitationClick}
-            className="prose prose-sm dark:prose-invert max-w-none prose-a:text-link prose-a:underline"
-          />
-        ) : isStreaming ? (
-          <div className="flex items-center gap-1.5 text-muted-foreground">
-            <Loader2 size={12} className="animate-spin" />
-            <span className="text-xs">Thinking...</span>
-          </div>
-        ) : null}
-
-        {message.sources && message.sources.length > 0 && (
-          <div className="border-t border-border pt-1.5 mt-1.5">
-            <p className="text-xs text-muted-foreground mb-1">
-              {message.sources.length} source{message.sources.length !== 1 ? "s" : ""}
-            </p>
-            <div className="space-y-0.5">
-              {message.sources.slice(0, 3).map((s) => (
-                <button
-                  key={s.chunk_id}
-                  type="button"
-                  onClick={() => scrollToTime(Math.floor(s.start_time))}
-                  className="block w-full text-left text-xs text-link hover:underline truncate"
-                >
-                  {s.timestamp} {s.speaker_label ? `(${s.speaker_label})` : ""} — {s.text.slice(0, 80)}...
-                </button>
-              ))}
-            </div>
-          </div>
-        )}
-      </div>
     </div>
   );
 }
