+++ apps/web/src/components/EpisodeChatExports.ts
@@ -1,0 +1,140 @@
+/**
+ * Conversation-export helpers for EpisodeChat (split out in #665).
+ * Generates Markdown and plain-text downloads from a Q&A message stream.
+ */
+import { formatTimestamp } from "@/lib/timestamp";
+import { formatDateTime } from "@/lib/dateFormat";
+import type { Source } from "@/lib/citations";
+import type { Message } from "./EpisodeChatMessage";
+
+export interface ConversationMeta {
+  feedTitle?: string | null;
+  episodeTitle: string;
+  episodeDescription?: string | null;
+}
+
+export function stripHtml(html: string): string {
+  return html.replace(/<[^>]*>/g, "").replace(/\s+/g, " ").trim();
+}
+
+export function downloadFile(
+  content: string,
+  filename: string,
+  mimeType: string,
+) {
+  const blob = new Blob([content], { type: mimeType });
+  const url = URL.createObjectURL(blob);
+  const a = document.createElement("a");
+  a.href = url;
+  a.download = filename;
+  document.body.appendChild(a);
+  a.click();
+  document.body.removeChild(a);
+  URL.revokeObjectURL(url);
+}
+
+export function pairMessages(
+  messages: Message[],
+): { question: string; answer: string; sources: Source[] }[] {
+  const pairs: { question: string; answer: string; sources: Source[] }[] = [];
+  for (let i = 0; i < messages.length; i++) {
+    const m = messages[i];
+    if (m.role !== "user") continue;
+    const next = messages[i + 1];
+    if (next && next.role === "assistant" && next.content) {
+      pairs.push({
+        question: m.content,
+        answer: next.content,
+        sources: next.sources ?? [],
+      });
+    }
+  }
+  return pairs;
+}
+
+export function generateConversationMarkdown(
+  meta: ConversationMeta,
+  messages: Message[],
+): string {
+  const lines: string[] = [];
+  lines.push(`# ${meta.feedTitle ?? "Podcast"}`);
+  lines.push("");
+  lines.push(`## ${meta.episodeTitle}`);
+  lines.push("");
+  if (meta.episodeDescription) {
+    lines.push(stripHtml(meta.episodeDescription));
+    lines.push("");
+  }
+  lines.push(`*Exported: ${formatDateTime(new Date())}*`);
+  lines.push("");
+  lines.push("---");
+  lines.push("");
+
+  const pairs = pairMessages(messages);
+  pairs.forEach((pair, idx) => {
+    lines.push(`### Q${idx + 1}: ${pair.question}`);
+    lines.push("");
+    lines.push(pair.answer);
+    lines.push("");
+    if (pair.sources.length > 0) {
+      lines.push(`**Sources (${pair.sources.length}):**`);
+      lines.push("");
+      for (const s of pair.sources) {
+        const speaker = s.speaker_label ? ` **${s.speaker_label}**` : "";
+        lines.push(
+          `- \`${s.timestamp ?? formatTimestamp(s.start_time)}\`${speaker} — ${s.text}`,
+        );
+      }
+      lines.push("");
+    }
+    lines.push("---");
+    lines.push("");
+  });
+
+  lines.push("*Generated by Podlog*");
+  return lines.join("\n");
+}
+
+export function generateConversationText(
+  meta: ConversationMeta,
+  messages: Message[],
+): string {
+  const lines: string[] = [];
+  lines.push(meta.feedTitle ?? "Podcast");
+  lines.push("=".repeat((meta.feedTitle ?? "Podcast").length));
+  lines.push("");
+  lines.push(meta.episodeTitle);
+  lines.push("-".repeat(meta.episodeTitle.length));
+  lines.push("");
+  if (meta.episodeDescription) {
+    lines.push(stripHtml(meta.episodeDescription));
+    lines.push("");
+  }
+  lines.push(`Exported: ${formatDateTime(new Date())}`);
+  lines.push("");
+  lines.push("========================================");
+  lines.push("");
+
+  const pairs = pairMessages(messages);
+  pairs.forEach((pair, idx) => {
+    lines.push(`Q${idx + 1}: ${pair.question}`);
+    lines.push("");
+    lines.push(`A${idx + 1}: ${pair.answer}`);
+    lines.push("");
+    if (pair.sources.length > 0) {
+      lines.push(`Sources (${pair.sources.length}):`);
+      for (const s of pair.sources) {
+        const speaker = s.speaker_label ? ` ${s.speaker_label}:` : "";
+        lines.push(
+          `  [${s.timestamp ?? formatTimestamp(s.start_time)}]${speaker} ${s.text}`,
+        );
+      }
+      lines.push("");
+    }
+    lines.push("----------------------------------------");
+    lines.push("");
+  });
+
+  lines.push("Generated by Podlog");
+  return lines.join("\n");
+}
